[ERROR] 사용자가 잘못된 값을 입력시 다시 입력할 수 있게 한다
