+++ src/Util.js
@@ -1,5 +1,29 @@
+import readline from "readline";
+
+function readLineAsync(query) {
+  return new Promise((resolve, reject) => {
+    if (arguments.length !== 1) {
+      reject(new Error("arguments must be 1"));
+    }
+
+    if (typeof query !== "string") {
+      reject(new Error("query must be string"));
+    }
+
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(query, (input) => {
+      rl.close();
+      resolve(input);
+    });
+  });
+}
+
 const getRandomNumber = () => {
   return Math.floor(Math.random() * 10);
 };
 
-export { getRandomNumber };
+export { getRandomNumber, readLineAsync };
